Moves assets and adds Next.js Link
