+++ src/components/GlobalStyle.js
@@ -5,7 +5,7 @@ const GlobalStyle = createGlobalStyle`
     box-sizing: border-box;
     margin: 0;
     padding: 0;
-    font-family: 'Roboto', sans-serif; 
+    font-family: "Roboto"; 
   }
 
   :root {
